Configure fife mock for tests in pytest
This means we can actually get rid of the run_tests.py wrapper, yay!
This also paves the way for running tests in parallel (which kinda works
already, except a few tests).

Added pytest-warnings plugin which will show warnings by default
(equivalent to the -Wd flag for python).
