+++ setup.cfg
@@ -2,9 +2,6 @@
 text_domains=[('unknown-horizons', 'po/uh/'), ('unknown-horizons-server', 'po/uh-server/')]
 bug_contact=team@unknown-horizons.org
 desktop_files=[('share/applications', 'po/uh/', ('content/packages/unknown-horizons.desktop.in',))]
-[nosetests]
-traverse-namespace=1
-attr=!long,!gui
 [pycodestyle]
 exclude=horizons/ext/
 count=1
